+++ Vistas/MenuAlumno.cs
@@ -14,10 +14,10 @@ namespace CAI_Intensivo2025_Grupo4.Vistas
 {
     public partial class MenuAlumno : Form
     {
-        private long _alumnoId;
-        private long _carreraId;
+        private int _alumnoId;
+        private int _carreraId;
 
-        public MenuAlumno(long alumnoId, long carreraId)
+        public MenuAlumno(int alumnoId, int carreraId)
         {
             InitializeComponent();
             _alumnoId = alumnoId;
@@ -30,7 +30,7 @@ namespace CAI_Intensivo2025_Grupo4.Vistas
         }
         private void btnInscripcionMaterias_Click(object sender, EventArgs e)
         {
-            InscripcionMaterias formMaterias = new InscripcionMaterias(_alumnoId, _carreraId);
+            InscripcionMaterias formMaterias = new InscripcionMaterias(_alumnoId,_carreraId);
             formMaterias.Show();
         }
